test(gateway): split node pairing reconnect coverage
