+++ src/gateway/server.node-pairing-reconnect.test.ts
@@ -1,0 +1,370 @@
+// Node pairing reconnect tests cover capability approval, metadata refreshes,
+// and reapproval when a paired node changes its declared surface.
+import { afterAll, beforeAll, describe, expect, test, vi } from "vitest";
+import { WebSocket } from "ws";
+import type { HelloOk } from "../../packages/gateway-protocol/src/index.js";
+import { getPairedDevice, listDevicePairing } from "../infra/device-pairing.js";
+import { approveNodePairing, listNodePairing, requestNodePairing } from "../infra/node-pairing.js";
+import { GATEWAY_CLIENT_MODES, GATEWAY_CLIENT_NAMES } from "../utils/message-channel.js";
+import { openTrackedWs, pairDeviceIdentity } from "./device-authz.test-helpers.js";
+import {
+  connectNodeClient,
+  createNodePairingTestState,
+  describeWithGatewayServer,
+  findPairedNode,
+  requireApprovedPairing,
+} from "./server.node-pairing.test-support.js";
+import { connectGatewayClient } from "./test-helpers.e2e.js";
+import {
+  connectOk,
+  installGatewayTestHooks,
+  rpcReq,
+  startServerWithClient,
+} from "./test-helpers.js";
+
+installGatewayTestHooks({ scope: "suite" });
+
+const { cleanup: cleanupNodePairingTestState, setup: setupNodePairingTestState } =
+  createNodePairingTestState("openclaw-node-pair-reconnect-");
+
+async function expectRePairingRequest(params: {
+  started: Awaited<ReturnType<typeof startServerWithClient>>;
+  pairedName: string;
+  initialCommands?: string[];
+  reconnectCommands: string[];
+  approvalScopes: string[];
+  expectedVisibleCommands: string[];
+}) {
+  const pairedNode = await pairDeviceIdentity({
+    name: params.pairedName,
+    role: "node",
+    scopes: [],
+    clientId: GATEWAY_CLIENT_NAMES.NODE_HOST,
+    clientMode: GATEWAY_CLIENT_MODES.NODE,
+  });
+
+  let controlWs: WebSocket | undefined;
+  let firstClient: Awaited<ReturnType<typeof connectNodeClient>> | undefined;
+  let nodeClient: Awaited<ReturnType<typeof connectNodeClient>> | undefined;
+  try {
+    controlWs = await openTrackedWs(params.started.port);
+    await connectOk(controlWs, { token: "secret" });
+
+    if (params.initialCommands) {
+      firstClient = await connectNodeClient({
+        port: params.started.port,
+        deviceIdentity: pairedNode.identity,
+        commands: params.initialCommands,
+      });
+      await firstClient.stopAndWait();
+    }
+
+    const request = await requestNodePairing({
+      nodeId: pairedNode.identity.deviceId,
+      platform: "macos",
+      deviceFamily: "Mac",
+      ...(params.initialCommands ? { commands: params.initialCommands } : {}),
+    });
+    await approveNodePairing(request.request.requestId, {
+      callerScopes: params.approvalScopes,
+    });
+
+    nodeClient = await connectNodeClient({
+      port: params.started.port,
+      deviceIdentity: pairedNode.identity,
+      commands: params.reconnectCommands,
+    });
+    const connectedControlWs = controlWs;
+
+    type NodeDiagnostics = {
+      nodeId: string;
+      connected?: boolean;
+      commands?: string[];
+      approvalState?: string;
+      pendingRequestId?: string;
+      pendingDeclaredCommands?: string[];
+    };
+    let lastNodes: NodeDiagnostics[] = [];
+    await vi.waitFor(async () => {
+      const list = await rpcReq<{
+        nodes?: NodeDiagnostics[];
+      }>(connectedControlWs, "node.list", {});
+      lastNodes = list.payload?.nodes ?? [];
+      const node = lastNodes.find(
+        (entry) => entry.nodeId === pairedNode.identity.deviceId && entry.connected,
+      );
+      if (
+        JSON.stringify(node?.commands?.toSorted() ?? []) ===
+        JSON.stringify(params.expectedVisibleCommands)
+      ) {
+        return;
+      }
+      throw new Error(`node commands not visible yet: ${JSON.stringify(lastNodes)}`);
+    });
+
+    expect(
+      lastNodes
+        .find((entry) => entry.nodeId === pairedNode.identity.deviceId && entry.connected)
+        ?.commands?.toSorted(),
+      JSON.stringify(lastNodes),
+    ).toEqual(params.expectedVisibleCommands);
+
+    const pairing = await listNodePairing();
+    const pending = pairing.pending?.find((entry) => entry.nodeId === pairedNode.identity.deviceId);
+    expect(pending?.nodeId).toBe(pairedNode.identity.deviceId);
+    expect(pending?.commands).toEqual(params.reconnectCommands);
+    const listedNode = lastNodes.find((entry) => entry.nodeId === pairedNode.identity.deviceId);
+    expect(listedNode).toMatchObject({
+      approvalState: "pending-reapproval",
+      pendingRequestId: pending?.requestId,
+      pendingDeclaredCommands: params.reconnectCommands,
+      commands: params.expectedVisibleCommands,
+    });
+
+    const described = await rpcReq<NodeDiagnostics>(connectedControlWs, "node.describe", {
+      nodeId: pairedNode.identity.deviceId,
+    });
+    expect(described.payload).toMatchObject({
+      approvalState: "pending-reapproval",
+      pendingRequestId: pending?.requestId,
+      pendingDeclaredCommands: params.reconnectCommands,
+      commands: params.expectedVisibleCommands,
+    });
+  } finally {
+    controlWs?.close();
+    await firstClient?.stopAndWait();
+    await nodeClient?.stopAndWait();
+  }
+}
+
+describe("gateway paired node reconnects", () => {
+  beforeAll(async () => {
+    await setupNodePairingTestState();
+  });
+
+  afterAll(async () => {
+    await cleanupNodePairingTestState();
+  });
+
+  describeWithGatewayServer("paired node reconnects", (getStarted) => {
+    test("withholds plugin surface URLs until the node capability is approved", async () => {
+      // The shared Gateway harness disables Canvas startup; expose its descriptor
+      // so this handshake test exercises production capability issuance.
+      const previousSkipCanvasHost = process.env.OPENCLAW_SKIP_CANVAS_HOST;
+      delete process.env.OPENCLAW_SKIP_CANVAS_HOST;
+      try {
+        const pairedNode = await pairDeviceIdentity({
+          name: "node-plugin-surface-approval",
+          role: "node",
+          scopes: [],
+          clientId: GATEWAY_CLIENT_NAMES.NODE_HOST,
+          clientMode: GATEWAY_CLIENT_MODES.NODE,
+        });
+        let pendingHello: HelloOk | undefined;
+        const pendingClient = await connectNodeClient({
+          port: getStarted().port,
+          deviceIdentity: pairedNode.identity,
+          caps: ["canvas"],
+          commands: [],
+          onHelloOk: (hello) => {
+            pendingHello = hello;
+          },
+        });
+        await pendingClient.stopAndWait();
+
+        expect(pendingHello?.pluginSurfaceUrls).toBeUndefined();
+        const pending = (await listNodePairing()).pending.find(
+          (entry) => entry.nodeId === pairedNode.identity.deviceId,
+        );
+        expect(pending?.caps).toEqual(["canvas"]);
+        requireApprovedPairing(
+          await approveNodePairing(pending?.requestId ?? "", {
+            callerScopes: ["operator.pairing"],
+          }),
+        );
+
+        let approvedHello: HelloOk | undefined;
+        const approvedClient = await connectNodeClient({
+          port: getStarted().port,
+          deviceIdentity: pairedNode.identity,
+          caps: ["canvas"],
+          commands: [],
+          onHelloOk: (hello) => {
+            approvedHello = hello;
+          },
+        });
+        try {
+          expect(approvedHello?.pluginSurfaceUrls?.canvas).toMatch(
+            /^http:\/\/127\.0\.0\.1:\d+\/__openclaw__\/cap\/[^/]+$/,
+          );
+        } finally {
+          await approvedClient.stopAndWait();
+        }
+      } finally {
+        if (previousSkipCanvasHost === undefined) {
+          delete process.env.OPENCLAW_SKIP_CANVAS_HOST;
+        } else {
+          process.env.OPENCLAW_SKIP_CANVAS_HOST = previousSkipCanvasHost;
+        }
+      }
+    });
+
+    test("keeps iOS approval when a transient permission becomes unavailable", async () => {
+      const pairedNode = await pairDeviceIdentity({
+        name: "ios-transient-permission",
+        role: "node",
+        scopes: [],
+        clientId: GATEWAY_CLIENT_NAMES.IOS_APP,
+        clientMode: GATEWAY_CLIENT_MODES.NODE,
+      });
+      const initialPermissions = { camera: true, watchReachable: true };
+      const initial = await requestNodePairing({
+        nodeId: pairedNode.identity.deviceId,
+        platform: "ios",
+        deviceFamily: "iPhone",
+        commands: [],
+        permissions: initialPermissions,
+      });
+      await approveNodePairing(initial.request.requestId, {
+        callerScopes: ["operator.pairing", "operator.write"],
+      });
+
+      const nodeClient = await connectGatewayClient({
+        url: `ws://127.0.0.1:${getStarted().port}`,
+        token: "secret",
+        role: "node",
+        clientName: GATEWAY_CLIENT_NAMES.IOS_APP,
+        clientDisplayName: "iPhone",
+        clientVersion: "1.0.0",
+        platform: "ios",
+        deviceFamily: "iPhone",
+        mode: GATEWAY_CLIENT_MODES.NODE,
+        scopes: [],
+        commands: [],
+        permissions: { camera: true, watchReachable: false },
+        deviceIdentity: pairedNode.identity,
+      });
+      await nodeClient.stopAndWait();
+
+      const pairing = await listNodePairing();
+      expect(pairing.pending.some((entry) => entry.nodeId === pairedNode.identity.deviceId)).toBe(
+        false,
+      );
+      await expect(findPairedNode(pairedNode.identity.deviceId)).resolves.toMatchObject({
+        permissions: initialPermissions,
+      });
+    });
+
+    test("clears stale reapproval when a node returns to its approved surface", async () => {
+      const pairedNode = await pairDeviceIdentity({
+        name: "node-reverted-reapproval",
+        role: "node",
+        scopes: [],
+        clientId: GATEWAY_CLIENT_NAMES.NODE_HOST,
+        clientMode: GATEWAY_CLIENT_MODES.NODE,
+      });
+      const initial = await requestNodePairing({
+        nodeId: pairedNode.identity.deviceId,
+        platform: "macos",
+        deviceFamily: "Mac",
+        commands: ["screen.snapshot"],
+      });
+      await approveNodePairing(initial.request.requestId, {
+        callerScopes: ["operator.pairing", "operator.write"],
+      });
+
+      const upgraded = await connectNodeClient({
+        port: getStarted().port,
+        deviceIdentity: pairedNode.identity,
+        commands: ["screen.snapshot", "system.run"],
+      });
+      await upgraded.stopAndWait();
+      expect(
+        (await listNodePairing()).pending.some(
+          (entry) => entry.nodeId === pairedNode.identity.deviceId,
+        ),
+      ).toBe(true);
+
+      const reverted = await connectNodeClient({
+        port: getStarted().port,
+        deviceIdentity: pairedNode.identity,
+        commands: ["screen.snapshot"],
+      });
+      await reverted.stopAndWait();
+
+      await vi.waitFor(async () => {
+        expect(
+          (await listNodePairing()).pending.some(
+            (entry) => entry.nodeId === pairedNode.identity.deviceId,
+          ),
+        ).toBe(false);
+      });
+    });
+
+    test("refreshes a paired macOS app node version without a repair request", async () => {
+      const started = getStarted();
+      const pairedNode = await pairDeviceIdentity({
+        name: "macos-version-refresh",
+        role: "node",
+        scopes: [],
+        clientId: GATEWAY_CLIENT_NAMES.MACOS_APP,
+        clientMode: GATEWAY_CLIENT_MODES.NODE,
+        platform: "macOS 26.5.1",
+        deviceFamily: "Mac",
+      });
+      const nodeRequest = await requestNodePairing({
+        nodeId: pairedNode.identity.deviceId,
+        platform: "macOS 26.5.1",
+        deviceFamily: "Mac",
+        commands: ["system.info"],
+      });
+      requireApprovedPairing(
+        await approveNodePairing(nodeRequest.request.requestId, {
+          callerScopes: ["operator.pairing", "operator.write", "operator.admin"],
+        }),
+      );
+
+      const nodeClient = await connectNodeClient({
+        port: started.port,
+        deviceIdentity: pairedNode.identity,
+        commands: ["system.info"],
+        clientName: GATEWAY_CLIENT_NAMES.MACOS_APP,
+        platform: "macOS 26.5.2",
+        deviceFamily: "Mac",
+      });
+      try {
+        await vi.waitFor(async () => {
+          const pairedDevice = await getPairedDevice(pairedNode.identity.deviceId);
+          expect(pairedDevice?.platform).toBe("macOS 26.5.2");
+        });
+        const devicePairing = await listDevicePairing();
+        expect(
+          devicePairing.pending.find((entry) => entry.deviceId === pairedNode.identity.deviceId),
+        ).toBeUndefined();
+      } finally {
+        await nodeClient.stopAndWait();
+      }
+    });
+
+    test("requests re-pairing when a paired node reconnects with upgraded commands", async () => {
+      await expectRePairingRequest({
+        started: getStarted(),
+        pairedName: "node-command-pin",
+        initialCommands: ["screen.snapshot"],
+        reconnectCommands: ["screen.snapshot", "system.run"],
+        approvalScopes: ["operator.pairing", "operator.write"],
+        expectedVisibleCommands: ["screen.snapshot"],
+      });
+    });
+
+    test("requests re-pairing when a commandless paired node reconnects with system.run", async () => {
+      await expectRePairingRequest({
+        started: getStarted(),
+        pairedName: "node-command-empty",
+        reconnectCommands: ["screen.snapshot", "system.run"],
+        approvalScopes: ["operator.pairing"],
+        expectedVisibleCommands: [],
+      });
+    });
+  });
+});
